refactor(storage): move storage from types package
