+++ pkg/storage/storage.go
@@ -1,10 +1,14 @@
-package types
+package storage
 
-import "io"
+import (
+	"io"
+
+	"github.com/fmich7/fyle/pkg/types"
+)
 
 // Storage represents a storage interface for server
 type Storage interface {
-	StoreFile(file *File) error
+	StoreFile(file *types.File) error
 	RetrieveFile(path string) (io.ReadCloser, error)
 	GetFileUploadsLocation() string
 }
